refactor(Banner): rename into BannerTags, cleanup
