+++ components/Banner/ScrollAnimation/useScrollDirection.ts
